agent/requirements.txt
python-socketio
requests
websocket-client
pymongo[srv]
dotenv
pydantic

langgraph
langchain
langchain-google-genai

schedule